+++ osu.Game/Screens/Edit/Timing/EffectSection.cs
@@ -19,7 +19,7 @@ namespace osu.Game.Screens.Edit.Timing
             Flow.AddRange(new[]
             {
                 kiai = new LabelledSwitchButton { Label = "歌曲高潮时间" },
-                omitBarLine = new LabelledSwitchButton { Label = "Skip Bar Line" },
+                omitBarLine = new LabelledSwitchButton { Label = "越过第一条横线(?)" },
             });
         }